+++ app/src/main/java/com/karrar/weather_app/data/domain/Daily.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Daily(
     @SerializedName("clouds")
-    val clouds: Int?,
+    val clouds: Double?,
     @SerializedName("dew_point")
     val dewPoint: Double?,
     @SerializedName("dt")
@@ -13,7 +13,7 @@ data class Daily(
     @SerializedName("feels_like")
     val feelsLike: FeelsLike?,
     @SerializedName("humidity")
-    val humidity: Int?,
+    val humidity: Double?,
     @SerializedName("moon_phase")
     val moonPhase: Double?,
     @SerializedName("moonrise")
@@ -23,7 +23,7 @@ data class Daily(
     @SerializedName("pop")
     val pop: Double?,
     @SerializedName("pressure")
-    val pressure: Int?,
+    val pressure: Double?,
     @SerializedName("rain")
     val rain: Double?,
     @SerializedName("sunrise")
@@ -37,7 +37,7 @@ data class Daily(
     @SerializedName("weather")
     val weatherStatus: List<WeatherStatus>?,
     @SerializedName("wind_deg")
-    val windDeg: Int?,
+    val windDeg: Double?,
     @SerializedName("wind_gust")
     val windGust: Double?,
     @SerializedName("wind_speed")
