+++ app/src/main/java/com/karrar/weather_app/data/domain/Minutely.kt
@@ -7,5 +7,5 @@ data class Minutely(
     @SerializedName("dt")
     val dt: Int?,
     @SerializedName("precipitation")
-    val precipitation: Int?
+    val precipitation: Double?
 )
